Changed placement of vpc stack, switched to single quotes, and renamed validation file
